Able to sort multiple lightpack modules
If you have more than 2 lightpack modules, it may sometimes read
each module in a different order and therefore messing around
with the LEDs positions that the users sets. You can now change
the order of the modules in the connections tab so that LEDs
positions matchies the intended positions without needing to
rearrange every LED in the position tab.