On first start with filter, it will show correct number of leds
Since the filter has no settings file to read, it will not show
anything till it has connected to the gui. Once connected, it will
set the filter's default position of the number leds that are
connected.

This also fixes the buffer overflow issue with the sockets which
now requires a buffer size to be passed.