Added SSE2 implementation for NV12 average pixels
This is the fastest method so far to convert any YUV compressed
video stream. The routine is 4 times faster than the simple
software implementation that was currently used and therefore
seeing improvements of 1-3% less CPU overhead with just MADVR
and 5% less with SVP compared to before.

The routine uses SSE2 which is supported all the way back to
Pentium 4 and falls back to simple software solution if cannot
support SSE2.

Raw RGB32 is still the slowest method.