Fixes build scripts for app-mutex and lightpack node modules
The scripts now share a single python script which is called targeted to
or from the working directory with a 'binding.gyp' file. This uses the
compilation from nw-gyp instead of compiling with visual studio directly.
This allows the node modules to work on other PCs that failed before.

You must also compile lightpack.lib with the static runtime (MT) and not
the default dynamic runtime (MD).