Fixes lights displaying from filter when installed
Previously the filter was reading the wrong path of node webkit's save
path when installed. Since that save path may change over time, the save
path is queried from node webkit and is sent to the filter everytime the
filter connects. As long as the filter reads settings once, it will not
do it again.