Add a delay to automate events and refactor
Added the delay because the automated events that may fire after one
another with small time frame may not run the second event (in terms on
the online services such as IFTTT). You can add a delay for multiple
quick events so that the second event (or last) may always fire
successfully.

Also refactored the automation code to its own file.