Fixes the freezing during playback from reconnection
Hidapi was taking about 5sec sometimes to re-enumerate the avaliable usb
devices which happens on the rendering thread causing it to freeze. It
is now moved to a separate thread. Meanwhile when it does freeze, skip
showing the lights (so they do not get blocked by the critial section)
so it does not lag behind the video. Finally update Hidapi's version
which actually may fix all the freezing.